refactor #1 reduxフォルダーの作成
