+++ src/redux/store.ts
@@ -1,0 +1,9 @@
+import { configureStore, createSlice } from "@reduxjs/toolkit";
+import { counterSlice } from "./counterSlice";
+
+export const store = configureStore({
+    reducer: {
+      counter: counterSlice.reducer,
+      ideaList: counterSlice.reducer,
+    },
+});
